Split chrome decision into ToolbarVisibilityDecision + AITabChromeDecision twins
